extract an abstract model for sparse and dense sentence transformer translator
