+++ ml-algorithms/src/main/java/org/opensearch/ml/engine/algorithms/sparse_encoding/SparseEncodingTranslator.java
@@ -19,6 +19,7 @@ import ai.djl.translate.TranslatorContext;
 import org.opensearch.ml.common.output.model.MLResultDataType;
 import org.opensearch.ml.common.output.model.ModelTensor;
 import org.opensearch.ml.common.output.model.ModelTensors;
+import org.opensearch.ml.engine.algorithms.SentenceTransformerTranslator;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -27,38 +28,7 @@ import java.util.*;
 
 import static org.opensearch.ml.common.CommonValue.ML_MAP_RESPONSE_KEY;
 
-public class SparseEncodingTranslator implements ServingTranslator {
-    private HuggingFaceTokenizer tokenizer;
-
-    @Override
-    public Batchifier getBatchifier() {
-        return Batchifier.STACK;
-    }
-    @Override
-    public void prepare(TranslatorContext ctx) throws IOException {
-        Path path = ctx.getModel().getModelPath();
-        tokenizer = HuggingFaceTokenizer.builder().optPadding(true).optTokenizerPath(path.resolve("tokenizer.json")).build();
-    }
-
-    @Override
-    public NDList processInput(TranslatorContext ctx, Input input) {
-        String sentence = input.getAsString(0);
-        NDManager manager = ctx.getNDManager();
-        NDList ndList = new NDList();
-        Encoding encodings = tokenizer.encode(sentence);
-        long[] indices = encodings.getIds();
-        long[] attentionMask = encodings.getAttentionMask();
-
-        NDArray indicesArray = manager.create(indices);
-        indicesArray.setName("input1.input_ids");
-
-        NDArray attentionMaskArray = manager.create(attentionMask);
-        attentionMaskArray.setName("input1.attention_mask");
-
-        ndList.add(indicesArray);
-        ndList.add(attentionMaskArray);
-        return ndList;
-    }
+public class SparseEncodingTranslator extends SentenceTransformerTranslator {
     private Map<String, Float>  convertOutput(NDArray array)
     {
         Map<String, Float> map = new HashMap<>();
@@ -93,9 +63,5 @@ public class SparseEncodingTranslator implements ServingTranslator {
         ModelTensors modelTensorOutput = new ModelTensors(outputs);
         output.add(modelTensorOutput.toBytes());
         return output;
-    }
-
-    @Override
-    public void setArguments(Map<String, ?> arguments) {
     }
 }
